move 'or' tests to separate file
