+++ types/ramda/tsconfig.json
@@ -158,6 +158,7 @@
         "test/objOf-tests.ts",
         "test/of-tests.ts",
         "test/omit-tests.ts",
-        "test/once-tests.ts"
+        "test/once-tests.ts",
+        "test/or-tests.ts"
     ]
 }
